pyproject.toml
[project]
name = "mock-interviewing-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pydantic-ai>=1.0.7",
    "pydantic-graph>=1.0.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
]
